requirements.txt
streamlit==1.31.1
pdf2image==1.17.0
Pillow==10.2.0
pandas==2.2.1
numpy==1.26.4 